feature: vite testing
