+++ tensormap-client/jest.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+    testEnvironment: "jsdom",
+    moduleNameMapper: {
+      "^.+\\.svg$": "jest-svg-transformer",
+      "^.+\\.(css|less|scss)$": "identity-obj-proxy"
+    },
+    setupFilesAfterEnv: [
+      "<rootDir>/src/setupTests.js"
+    ]
+};
