+++ tensormap-client/package.json
@@ -3,44 +3,26 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emotion/core": "^11.0.0",
-    "@emotion/react": "^11.4.0",
-    "@emotion/styled": "^11.3.0",
-    "@projectstorm/react-diagrams": "^6.5.2",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
     "@vitejs/plugin-react": "^4.2.1",
     "axios": "^0.21.1",
     "closest": "0.0.1",
-    "dagre": "^0.8.5",
-    "lodash": "^4.17.21",
-    "mathjs": "^9.4.4",
-    "pathfinding": "^0.4.18",
-    "paths-js": "^0.4.11",
-    "prop-types": "^15.8.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-heatmap-grid": "^0.9.0",
     "react-redux": "^7.2.4",
     "react-router": "^5.2.0",
     "react-router-dom": "^6.21.3",
     "reactflow": "^11.7.2",
     "recoil": "^0.7.7",
-    "redux": "^4.1.0",
-    "resize-observer-polyfill": "^1.5.1",
-    "rxjs": "^7.2.0",
     "semantic-ui-css": "^2.4.1",
     "semantic-ui-react": "^2.0.3",
     "socket.io-client": "^4.1.3",
-    "typescript": "^4.3.5",
-    "web-vitals": "^1.1.2",
-    "vite": "^5.0.8"
+    "web-vitals": "^1.1.2"
   },
   "scripts": {
     "start": "vite",
     "build": "vite build",
-    "test": "react-scripts test",
+    "test": "jest",
     "eject": "react-scripts eject",
     "lint": "eslint .",
     "lint:fix": "eslint --fix ."
@@ -64,12 +46,23 @@
     ]
   },
   "devDependencies": {
+    "@babel/core": "^7.24.3",
+    "@babel/preset-env": "^7.24.3",
+    "@babel/preset-react": "^7.24.1",
+    "@testing-library/jest-dom": "^6.4.2",
+    "@testing-library/react": "^14.2.2",
     "@types/lodash": "^4.14.171",
+    "babel-jest": "^29.7.0",
     "eslint": "^8.47.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-plugin-import": "^2.28.0",
     "eslint-plugin-jsx-a11y": "^6.7.1",
     "eslint-plugin-react": "^7.33.2",
-    "eslint-plugin-react-hooks": "^4.6.0"
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "identity-obj-proxy": "^3.0.0",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
+    "jest-svg-transformer": "^1.0.0",
+    "vite": "^5.2.2"
   }
 }
