+++ tensormap-client/src/components/__tests__/NavBar.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import NavBar from '../Layout/NavBar/NavBar';
-import '@testing-library/jest-dom/extend-expect';
+import '@testing-library/jest-dom'
 
 describe('<NavBar />', () => {
   it('renders without crashing', () => {
