+++ tensormap-client/vite.config.js
